# target/classes/openapi.yaml
openapi: 3.0.1
info:
  description: This specification is intended to describe the functionality of the
    photo album application
  title: Photo album OpenAPI specification
  version: 0.0.1
servers:
- description: Dev server
  url: http://localhost:8080
paths:
...
  /pages:
    post:
      operationId: createPage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageCreateDto'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageReadDto'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Method for created page
      tags:
      - PageController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: PageController
  /pages/{id}:
    delete:
      description: Method for delete page
      operationId: deletePage
      parameters:
      - description: ID for delete page
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Delete page by ID
      tags:
      - PageController
      x-accepts: application/json
      x-tags:
      - tag: PageController
  /pages/all/{albumId}:
    get:
      description: Find pages by album ID
      operationId: getAllPagesByAlbumId
      parameters:
      - description: "album id , which need to find pages"
        explode: false
        in: path
        name: albumId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pages'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Find all pages by album ID
      tags:
      - PageController
      x-accepts: application/json
      x-tags:
      - tag: PageController
  /pages/one/{id}:
    get:
      description: Find page by ID
      operationId: getPage
      parameters:
      - description: "id , which need to find page"
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageReadDto'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Find page by ID
      tags:
      - PageController
      x-accepts: application/json
      x-tags:
      - tag: PageController
  /pages/update/{id}:
    put:
      description: Method for update page
      operationId: updatePage
      parameters:
      - description: id that need to update page
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageCreateDto'
        description: Update an existent page in the album
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageReadDto'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Update cover
      tags:
      - PageController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: PageController
...
components:
  schemas:
...
    PageCreateDto:
      example:
        color: red
        tag_people: "Andrew,Egor"
        template_page: Horizontal image and text field below
        description: some words about page
        tag_place: "Germany,Berlin"
        animation: slide
      properties:
        description:
          example: some words about page
          type: string
        color:
          example: red
          type: string
        template_page:
          example: Horizontal image and text field below
          type: string
        tag_people:
          example: "Andrew,Egor"
          type: string
        tag_place:
          example: "Germany,Berlin"
          type: string
        animation:
          example: slide
          type: string
      type: object
    PageReadDto:
      example:
        color: red
        tag_people: "Andrew,Egor"
        album: null
        template_page: Horizontal image and text field below
        description: some words about page
        id: 11
        tag_place: "Germany,Berlin"
        animation: slide
      properties:
        id:
          example: 11
          format: int64
          type: integer
        description:
          example: some words about page
          type: string
        color:
          example: red
          type: string
        template_page:
          example: Horizontal image and text field below
          type: string
        tag_people:
          example: "Andrew,Egor"
          type: string
        tag_place:
          example: "Germany,Berlin"
          type: string
        animation:
          example: slide
          type: string
        album:
          $ref: '#/components/schemas/AlbumReadDto'
      type: object
...
    AlbumReadDto:
      example:
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 0
        page:
        - color: red
          tag_people: "Andrew,Egor"
          album: null
          template_page: Horizontal image and text field below
          description: some words about page
          id: 11
          tag_place: "Germany,Berlin"
          animation: slide
        - color: red
          tag_people: "Andrew,Egor"
          album: null
          template_page: Horizontal image and text field below
          description: some words about page
          id: 11
          tag_place: "Germany,Berlin"
          animation: slide
      properties:
        id:
          format: int64
          type: integer
        created_at:
          format: date-time
          type: string
        page:
          items:
            $ref: '#/components/schemas/PageReadDto'
          type: array
      type: object
...
    Pages:
      items:
        $ref: '#/components/schemas/PageReadDto'
      type: array
    Error:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object